ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4